Working insertion sort, rather optimized.
